feat: Added field validation contraints in the client, project and room dto classes.
- Changed the budget and price fields to the correct value BigDecimal for precise floating points.

- Changed quantity value to Integer, for future calculations.

+++ src/main/java/com/interiordesignplanner/project/ProjectDTO.java
@@ -1,5 +1,6 @@
 package com.interiordesignplanner.project;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.time.LocalDate;
 
@@ -30,7 +31,7 @@ public class ProjectDTO {
     private ProjectStatus status;
 
     // Estimated cost of project
-    private Integer budget;
+    private BigDecimal budget;
 
     // The start date and due date to help with planning and progress tracking
     private LocalDate startDate;
